add prettier-ignore
