+++ test/commands/schema-types.spec.ts
@@ -50,15 +50,23 @@ it("should successfully create d.ts files from the replicant schemas and create 
 	const outputPath = "./src/types/schemas/example.d.ts";
 	expect(fs.existsSync(outputPath)).toBe(true);
 
-	await expect(fs.readFileSync(outputPath, "utf8")).toMatchFileSnapshot(
-		"../fixtures/results/schema-types/example.d.ts",
-	);
+	expect(fs.readFileSync(outputPath, "utf8")).toMatchInlineSnapshot(`
+		"/* prettier-ignore */
+		/* eslint-disable */
+		/**
+		 * This file was automatically generated by json-schema-to-typescript.
+		 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
+		 * and run json-schema-to-typescript to regenerate this file.
+		 */
 
-	const indexPath = "./src/types/schemas/index.d.ts";
-	expect(fs.existsSync(indexPath)).toBe(true);
-	await expect(fs.readFileSync(indexPath, "utf8")).toMatchFileSnapshot(
-		"../fixtures/results/schema-types/index.d.ts",
-	);
+		export interface Example {
+			ip: string;
+			port: number;
+			password: string;
+			status: 'connected' | 'connecting' | 'disconnected' | 'error';
+		}
+		"
+	`);
 });
 
 it("should print an error when the target bundle does not have a schemas dir", async () => {
@@ -83,14 +91,21 @@ it("should successfully compile the config schema", async () => {
 	const outputPath = "./src/types/schemas/configschema.d.ts";
 	expect(fs.existsSync(outputPath)).toBe(true);
 
-	await expect(fs.readFileSync(outputPath, "utf8")).toMatchFileSnapshot(
-		"../fixtures/results/schema-types/configschema.d.ts",
-	);
-	expect(fs.readFileSync("./src/types/schemas/index.d.ts", "utf8"))
-		.toMatchInlineSnapshot(`
-		"/* eslint-disable */
-		// @ts-ignore
-		export * from './configschema';
+	expect(fs.readFileSync(outputPath, "utf8")).toMatchInlineSnapshot(`
+		"/* prettier-ignore */
+		/* eslint-disable */
+		/**
+		 * This file was automatically generated by json-schema-to-typescript.
+		 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
+		 * and run json-schema-to-typescript to regenerate this file.
+		 */
+
+		export interface Configschema {
+			username?: string;
+			value?: number;
+			nodefault?: string;
+			[k: string]: unknown;
+		}
 		"
 	`);
 });
