--- test/fixtures/results/schema-types/configschema.d.ts
@@ -1,13 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-export interface Configschema {
-	username?: string;
-	value?: number;
-	nodefault?: string;
-	[k: string]: unknown;
-}
