--- test/fixtures/results/schema-types/index.d.ts
@@ -1,3 +1,0 @@
-/* eslint-disable */
-// @ts-ignore
-export * from './example';
